move point and ray to separate headers
